+++ ext/CategoricalInterventionsComponentArraysExt.jl
@@ -4,7 +4,7 @@ using ComponentArrays
 using CategoricalInterventions
 const CI = CategoricalInterventions
 
-function CI.extend_parameters(::Model, p0::ComponentArray, additions::Dict{Symbol,<:Real})
+function CI.extend_parameters(::Model, p0::ComponentArray, additions::AbstractVector{<:Pair})
     names = collect(keys(p0))
     pairs = Any[n => p0[n] for n in names]
     append!(pairs, [k => v for (k, v) in additions])