+++ ext/CategoricalInterventionsLabelledArraysExt.jl
@@ -4,7 +4,7 @@ using LabelledArrays
 using CategoricalInterventions
 const CI = CategoricalInterventions
 
-function CI.extend_parameters(::Model, p0::LArray, additions::Dict{Symbol,<:Real})
+function CI.extend_parameters(::Model, p0::LArray, additions::AbstractVector{<:Pair})
     names = collect(propertynames(p0))
     vals = Any[p0[n] for n in names]
     for (k, v) in additions